+++ src/circuit/standard_gate_ops.rs
@@ -15,7 +15,7 @@
 
 use crate::states::{ProductState, Qubit, SuperPosition};
 use crate::Complex;
-use crate::{complex, complex_Im, complex_Im_array, complex_Re, complex_Re_array, COMPLEX_ZERO};
+use crate::{complex, complex_im, complex_im_array, complex_re, complex_re_array, COMPLEX_ZERO};
 use std::f64::consts::FRAC_1_SQRT_2;
 use std::ops::{Div, Mul};
 
@@ -36,23 +36,23 @@ use std::ops::{Div, Mul};
 #[rustfmt::skip]
 pub fn identity(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [complex_Re!(1f64), COMPLEX_ZERO],
-        Qubit::One =>  [COMPLEX_ZERO, complex_Re!(1f64)],
+        Qubit::Zero => [complex_re!(1f64), COMPLEX_ZERO],
+        Qubit::One =>  [COMPLEX_ZERO, complex_re!(1f64)],
     })
 }
 
 #[rustfmt::skip]
 pub fn hadamard(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => complex_Re_array!(FRAC_1_SQRT_2, FRAC_1_SQRT_2),
-        Qubit::One => complex_Re_array!(FRAC_1_SQRT_2, -FRAC_1_SQRT_2),
+        Qubit::Zero => complex_re_array!(FRAC_1_SQRT_2, FRAC_1_SQRT_2),
+        Qubit::One => complex_re_array!(FRAC_1_SQRT_2, -FRAC_1_SQRT_2),
     })
 }
 
 #[rustfmt::skip]
 pub fn rx(register: Qubit, angle: f64) -> SuperPosition {
-    let real_parts: Complex<f64> = complex_Re!((0.5f64.mul(angle)).cos());
-    let imaginary_part: Complex<f64> = complex_Im!(-(0.5f64.mul(angle)).sin());
+    let real_parts: Complex<f64> = complex_re!((0.5f64.mul(angle)).cos());
+    let imaginary_part: Complex<f64> = complex_im!(-(0.5f64.mul(angle)).sin());
     let zero_map: [Complex<f64>; 2] = [real_parts, imaginary_part];
     let one_map: [Complex<f64>; 2] = [imaginary_part, real_parts];
 
@@ -64,9 +64,9 @@ pub fn rx(register: Qubit, angle: f64) -> SuperPosition {
 
 #[rustfmt::skip]
 pub fn ry(register: Qubit, angle: f64) -> SuperPosition {
-    let cos_parts: Complex<f64> = complex_Re!((0.5f64.mul(angle)).cos());
-    let sin_part_pos: Complex<f64> = complex_Re!((0.5f64.mul(angle)).sin());
-    let sin_part_neg: Complex<f64> = complex_Re!(-(0.5f64.mul(angle)).sin());
+    let cos_parts: Complex<f64> = complex_re!((0.5f64.mul(angle)).cos());
+    let sin_part_pos: Complex<f64> = complex_re!((0.5f64.mul(angle)).sin());
+    let sin_part_neg: Complex<f64> = complex_re!(-(0.5f64.mul(angle)).sin());
     let zero_map: [Complex<f64>; 2] = [cos_parts, sin_part_pos];
     let one_map: [Complex<f64>; 2] = [sin_part_neg, cos_parts];
 
@@ -104,39 +104,39 @@ pub fn global_phase(register: Qubit, angle: f64) -> SuperPosition {
 #[rustfmt::skip]
 pub fn x90(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [COMPLEX_ZERO, complex_Im!(-1f64)],
-        Qubit::One => [complex_Im!(-1f64), COMPLEX_ZERO],
+        Qubit::Zero => [COMPLEX_ZERO, complex_im!(-1f64)],
+        Qubit::One => [complex_im!(-1f64), COMPLEX_ZERO],
     })
 }
 
 #[rustfmt::skip]
 pub fn y90(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [COMPLEX_ZERO, complex_Re!(-1f64)],
-        Qubit::One => [complex_Re!(1f64), COMPLEX_ZERO],
+        Qubit::Zero => [COMPLEX_ZERO, complex_re!(-1f64)],
+        Qubit::One => [complex_re!(1f64), COMPLEX_ZERO],
     })
 }
 
 #[rustfmt::skip]
 pub fn mx90(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [COMPLEX_ZERO, complex_Im!(1f64)],
-        Qubit::One => [complex_Im!(1f64), COMPLEX_ZERO],
+        Qubit::Zero => [COMPLEX_ZERO, complex_im!(1f64)],
+        Qubit::One => [complex_im!(1f64), COMPLEX_ZERO],
     })
 }
 
 #[rustfmt::skip]
 pub fn my90(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [COMPLEX_ZERO, complex_Re!(1f64)],
-        Qubit::One => [complex_Re!(-1f64), COMPLEX_ZERO],
+        Qubit::Zero => [COMPLEX_ZERO, complex_re!(1f64)],
+        Qubit::One => [complex_re!(-1f64), COMPLEX_ZERO],
     })
 }
 
 #[rustfmt::skip]
 pub fn tgate(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [complex_Re!(1f64), COMPLEX_ZERO],
+        Qubit::Zero => [complex_re!(1f64), COMPLEX_ZERO],
         Qubit::One => [COMPLEX_ZERO, complex!(FRAC_1_SQRT_2, FRAC_1_SQRT_2)],
     })
 }
@@ -144,7 +144,7 @@ pub fn tgate(register: Qubit) -> SuperPosition {
 #[rustfmt::skip]
 pub fn tgatedag(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [complex_Re!(1f64), COMPLEX_ZERO],
+        Qubit::Zero => [complex_re!(1f64), COMPLEX_ZERO],
         Qubit::One => [COMPLEX_ZERO, complex!(FRAC_1_SQRT_2, -FRAC_1_SQRT_2)],
     })
 }
@@ -152,40 +152,40 @@ pub fn tgatedag(register: Qubit) -> SuperPosition {
 #[rustfmt::skip]
 pub fn phase(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [complex_Re!(1f64), COMPLEX_ZERO],
-        Qubit::One => [COMPLEX_ZERO, complex_Im!(1f64)],
+        Qubit::Zero => [complex_re!(1f64), COMPLEX_ZERO],
+        Qubit::One => [COMPLEX_ZERO, complex_im!(1f64)],
     })
 }
 
 #[rustfmt::skip]
 pub fn phasedag(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [complex_Re!(1f64), COMPLEX_ZERO],
-        Qubit::One => [COMPLEX_ZERO, complex_Im!(-1f64)],
+        Qubit::Zero => [complex_re!(1f64), COMPLEX_ZERO],
+        Qubit::One => [COMPLEX_ZERO, complex_im!(-1f64)],
     })
 }
 
 #[rustfmt::skip]
 pub fn pauli_x(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [COMPLEX_ZERO, complex_Re!(1f64)],
-        Qubit::One => [complex_Re!(1f64), COMPLEX_ZERO],
+        Qubit::Zero => [COMPLEX_ZERO, complex_re!(1f64)],
+        Qubit::One => [complex_re!(1f64), COMPLEX_ZERO],
     })
 }
 
 #[rustfmt::skip]
 pub fn pauli_y(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [COMPLEX_ZERO, complex_Im!(1f64)],
-        Qubit::One => [complex_Im!(-1f64), COMPLEX_ZERO],
+        Qubit::Zero => [COMPLEX_ZERO, complex_im!(1f64)],
+        Qubit::One => [complex_im!(-1f64), COMPLEX_ZERO],
     })
 }
 
 #[rustfmt::skip]
 pub fn pauli_z(register: Qubit) -> SuperPosition {
     SuperPosition::new_with_register_unchecked::<2>(match register {
-        Qubit::Zero => [complex_Re!(1f64), COMPLEX_ZERO],
-        Qubit::One => [COMPLEX_ZERO, complex_Re!(-1f64)],
+        Qubit::Zero => [complex_re!(1f64), COMPLEX_ZERO],
+        Qubit::One => [COMPLEX_ZERO, complex_re!(-1f64)],
     })
 }
 
@@ -197,10 +197,10 @@ pub fn pauli_z(register: Qubit) -> SuperPosition {
 pub fn cnot(register: ProductState) -> SuperPosition {
     let input_register: [Qubit; 2] = [register.qubits[0], register.qubits[1]];
     SuperPosition::new_with_register_unchecked::<4>(match input_register {
-        [Qubit::Zero, Qubit::Zero] => complex_Re_array!(1f64, 0f64, 0f64, 0f64),
-        [Qubit::Zero, Qubit::One]  => complex_Re_array!(0f64, 1f64, 0f64, 0f64),
-        [Qubit::One, Qubit::Zero]  => complex_Re_array!(0f64, 0f64, 0f64, 1f64),
-        [Qubit::One, Qubit::One]   => complex_Re_array!(0f64, 0f64, 1f64, 0f64),
+        [Qubit::Zero, Qubit::Zero] => complex_re_array!(1f64, 0f64, 0f64, 0f64),
+        [Qubit::Zero, Qubit::One]  => complex_re_array!(0f64, 1f64, 0f64, 0f64),
+        [Qubit::One, Qubit::Zero]  => complex_re_array!(0f64, 0f64, 0f64, 1f64),
+        [Qubit::One, Qubit::One]   => complex_re_array!(0f64, 0f64, 1f64, 0f64),
     })
 }
 
@@ -208,10 +208,10 @@ pub fn cnot(register: ProductState) -> SuperPosition {
 pub fn cy(register: ProductState) -> SuperPosition {
     let input_register: [Qubit; 2] = [register.qubits[0], register.qubits[1]];
     SuperPosition::new_with_register_unchecked::<4>(match input_register {
-        [Qubit::Zero, Qubit::Zero] => complex_Re_array!(1f64, 0f64, 0f64, 0f64),
-        [Qubit::Zero, Qubit::One]  => complex_Re_array!(0f64, 1f64, 0f64, 0f64),
-        [Qubit::One, Qubit::Zero]  => complex_Im_array!(0f64, 0f64, 0f64, 1f64),
-        [Qubit::One, Qubit::One]   => complex_Im_array!(0f64, 0f64, -1f64, 0f64),
+        [Qubit::Zero, Qubit::Zero] => complex_re_array!(1f64, 0f64, 0f64, 0f64),
+        [Qubit::Zero, Qubit::One]  => complex_re_array!(0f64, 1f64, 0f64, 0f64),
+        [Qubit::One, Qubit::Zero]  => complex_im_array!(0f64, 0f64, 0f64, 1f64),
+        [Qubit::One, Qubit::One]   => complex_im_array!(0f64, 0f64, -1f64, 0f64),
     })
 }
 
@@ -219,10 +219,10 @@ pub fn cy(register: ProductState) -> SuperPosition {
 pub fn cz(register: ProductState) -> SuperPosition {
     let input_register: [Qubit; 2] = [register.qubits[0], register.qubits[1]];
     SuperPosition::new_with_register_unchecked::<4>(match input_register {
-        [Qubit::Zero, Qubit::Zero] => complex_Re_array!(1f64, 0f64, 0f64, 0f64),
-        [Qubit::Zero, Qubit::One]  => complex_Re_array!(0f64, 1f64, 0f64, 0f64),
-        [Qubit::One, Qubit::Zero]  => complex_Re_array!(0f64, 0f64, 1f64, 0f64),
-        [Qubit::One, Qubit::One]   => complex_Re_array!(0f64, 0f64, 0f64, -1f64),
+        [Qubit::Zero, Qubit::Zero] => complex_re_array!(1f64, 0f64, 0f64, 0f64),
+        [Qubit::Zero, Qubit::One]  => complex_re_array!(0f64, 1f64, 0f64, 0f64),
+        [Qubit::One, Qubit::Zero]  => complex_re_array!(0f64, 0f64, 1f64, 0f64),
+        [Qubit::One, Qubit::One]   => complex_re_array!(0f64, 0f64, 0f64, -1f64),
     })
 }
 
@@ -230,10 +230,10 @@ pub fn cz(register: ProductState) -> SuperPosition {
 pub fn swap(register: ProductState) -> SuperPosition {
     let input_register: [Qubit; 2] = [register.qubits[0], register.qubits[1]];
     SuperPosition::new_with_register_unchecked::<4>(match input_register {
-        [Qubit::Zero, Qubit::Zero] => complex_Re_array!(1f64, 0f64, 0f64, 0f64),
-        [Qubit::Zero, Qubit::One]  => complex_Re_array!(0f64, 0f64, 1f64, 0f64),
-        [Qubit::One, Qubit::Zero]  => complex_Re_array!(0f64, 1f64, 0f64, 0f64),
-        [Qubit::One, Qubit::One]   => complex_Re_array!(0f64, 0f64, 0f64, 1f64),
+        [Qubit::Zero, Qubit::Zero] => complex_re_array!(1f64, 0f64, 0f64, 0f64),
+        [Qubit::Zero, Qubit::One]  => complex_re_array!(0f64, 0f64, 1f64, 0f64),
+        [Qubit::One, Qubit::Zero]  => complex_re_array!(0f64, 1f64, 0f64, 0f64),
+        [Qubit::One, Qubit::One]   => complex_re_array!(0f64, 0f64, 0f64, 1f64),
     })
 }
 
@@ -242,9 +242,9 @@ pub fn cr(register: ProductState, angle: f64) -> SuperPosition {
     let input_register: [Qubit; 2] = [register.qubits[0], register.qubits[1]];
     let exp_array: [Complex<f64>; 4] = [COMPLEX_ZERO, COMPLEX_ZERO, COMPLEX_ZERO, Complex::<f64>::exp_im(angle)];
     SuperPosition::new_with_register_unchecked::<4>(match input_register {
-        [Qubit::Zero, Qubit::Zero] => complex_Re_array!(1f64, 0f64, 0f64, 0f64),
-        [Qubit::Zero, Qubit::One]  => complex_Re_array!(0f64, 1f64, 0f64, 0f64),
-        [Qubit::One, Qubit::Zero]  => complex_Re_array!(0f64, 0f64, 1f64, 0f64),
+        [Qubit::Zero, Qubit::Zero] => complex_re_array!(1f64, 0f64, 0f64, 0f64),
+        [Qubit::Zero, Qubit::One]  => complex_re_array!(0f64, 1f64, 0f64, 0f64),
+        [Qubit::One, Qubit::Zero]  => complex_re_array!(0f64, 0f64, 1f64, 0f64),
         [Qubit::One, Qubit::One]   => exp_array,
     })
 }
@@ -255,9 +255,9 @@ pub fn crk(register: ProductState, k: i32) -> SuperPosition {
     let exp_array: [Complex<f64>; 4] = 
         [COMPLEX_ZERO, COMPLEX_ZERO, COMPLEX_ZERO, Complex::<f64>::exp_im((2f64*std::f64::consts::PI).div(2f64.powi(k)))];
     SuperPosition::new_with_register_unchecked::<4>(match input_register {
-        [Qubit::Zero, Qubit::Zero] => complex_Re_array!(1f64, 0f64, 0f64, 0f64),
-        [Qubit::Zero, Qubit::One]  => complex_Re_array!(0f64, 1f64, 0f64, 0f64),
-        [Qubit::One, Qubit::Zero]  => complex_Re_array!(0f64, 0f64, 1f64, 0f64),
+        [Qubit::Zero, Qubit::Zero] => complex_re_array!(1f64, 0f64, 0f64, 0f64),
+        [Qubit::Zero, Qubit::One]  => complex_re_array!(0f64, 1f64, 0f64, 0f64),
+        [Qubit::One, Qubit::Zero]  => complex_re_array!(0f64, 0f64, 1f64, 0f64),
         [Qubit::One, Qubit::One]   => exp_array,
     })
 }
@@ -270,13 +270,13 @@ pub fn crk(register: ProductState, k: i32) -> SuperPosition {
 pub fn toffoli(register: ProductState) -> SuperPosition {
     let input_register: [Qubit; 3] = [register.qubits[0], register.qubits[1], register.qubits[2]];
     SuperPosition::new_with_register_unchecked::<8>(match input_register {
-            [Qubit::Zero, Qubit::Zero, Qubit::Zero] => {complex_Re_array!(1f64, 0f64, 0f64, 0f64, 0f64, 0f64, 0f64, 0f64) }
-            [Qubit::Zero, Qubit::Zero, Qubit::One] => { complex_Re_array!(0f64, 1f64, 0f64, 0f64, 0f64, 0f64, 0f64, 0f64) }
-            [Qubit::Zero, Qubit::One, Qubit::Zero] => { complex_Re_array!(0f64, 0f64, 1f64, 0f64, 0f64, 0f64, 0f64, 0f64) }
-            [Qubit::Zero, Qubit::One, Qubit::One] => {  complex_Re_array!(0f64, 0f64, 0f64, 1f64, 0f64, 0f64, 0f64, 0f64) }
-            [Qubit::One, Qubit::Zero, Qubit::Zero] => { complex_Re_array!(0f64, 0f64, 0f64, 0f64, 1f64, 0f64, 0f64, 0f64) }
-            [Qubit::One, Qubit::Zero, Qubit::One] => {  complex_Re_array!(0f64, 0f64, 0f64, 0f64, 0f64, 1f64, 0f64, 0f64) }
-            [Qubit::One, Qubit::One, Qubit::Zero] => {  complex_Re_array!(0f64, 0f64, 0f64, 0f64, 0f64, 0f64, 0f64, 1f64) }
-            [Qubit::One, Qubit::One, Qubit::One] => {   complex_Re_array!(0f64, 0f64, 0f64, 0f64, 0f64, 0f64, 1f64, 0f64) }
+            [Qubit::Zero, Qubit::Zero, Qubit::Zero] => { complex_re_array!(1f64, 0f64, 0f64, 0f64, 0f64, 0f64, 0f64, 0f64) }
+            [Qubit::Zero, Qubit::Zero, Qubit::One] => { complex_re_array!(0f64, 1f64, 0f64, 0f64, 0f64, 0f64, 0f64, 0f64) }
+            [Qubit::Zero, Qubit::One, Qubit::Zero] => { complex_re_array!(0f64, 0f64, 1f64, 0f64, 0f64, 0f64, 0f64, 0f64) }
+            [Qubit::Zero, Qubit::One, Qubit::One] => {  complex_re_array!(0f64, 0f64, 0f64, 1f64, 0f64, 0f64, 0f64, 0f64) }
+            [Qubit::One, Qubit::Zero, Qubit::Zero] => { complex_re_array!(0f64, 0f64, 0f64, 0f64, 1f64, 0f64, 0f64, 0f64) }
+            [Qubit::One, Qubit::Zero, Qubit::One] => {  complex_re_array!(0f64, 0f64, 0f64, 0f64, 0f64, 1f64, 0f64, 0f64) }
+            [Qubit::One, Qubit::One, Qubit::Zero] => {  complex_re_array!(0f64, 0f64, 0f64, 0f64, 0f64, 0f64, 0f64, 1f64) }
+            [Qubit::One, Qubit::One, Qubit::One] => {   complex_re_array!(0f64, 0f64, 0f64, 0f64, 0f64, 0f64, 1f64, 0f64) }
         })
 }